server/requirements.txt
Flask
python-dotenv
requests
google-auth
google-auth-oauthlib
google-auth-httplib2